pyproject.toml
[project]
name = "ai-claim-screener"
version = "0.1.0"
description = "ai insurance screener for level 1 fulfillment processing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.4",
    "botocore>=1.40.4",
    "cryptography>=45.0.6",
    "fastapi>=0.121.0",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.30",
    "langchain-community>=0.3.27",
    "mysql-connector-python>=9.5.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pymysql>=1.1.1",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "uvicorn>=0.38.0",
]
